Map C error to Python errors.
This provides Python Exception subclasses equivalent to each of the
errors defined by the libsensors API so that they can be individually
caught.  It also proivdes a function to handle raising the appropriate
error based on the return value of the function that caused it.

Tracebacks with this are not quite as nice as they were, adding an extra
function to the end, but there is unfortunately no way that I know of to
avoid this that is compatible with both Python 2 and Python 3.